+++ lib/app_manager.dart
@@ -1,8 +1,11 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_timezone/flutter_timezone.dart';
 import 'package:memorize_scripture/service_locator.dart';
 import 'package:memorize_scripture/services/data_repository/data_repository.dart';
+import 'package:memorize_scripture/services/notification_service.dart';
 import 'package:memorize_scripture/services/user_settings.dart';
 import 'package:timezone/data/latest_all.dart' as tz;
 import 'package:timezone/timezone.dart' as tz;
@@ -10,12 +13,14 @@ import 'package:timezone/timezone.dart' as tz;
 class AppManager {
   final themeListener = ValueNotifier<ThemeData>(_lightTheme);
   final userSettings = getIt<UserSettings>();
+  final notificationService = getIt<NotificationService>();
 
   Future<void> init() async {
     await userSettings.init();
     await _setDarkLightTheme();
     await getIt<DataRepository>().init();
-    await _initNotifications();
+    await getIt<NotificationService>().init();
+    await getIt<NotificationService>().scheduleNotifications();
   }
 
   Future<void> _setDarkLightTheme() async {
@@ -39,22 +44,3 @@ final _darkTheme = ThemeData(
   colorSchemeSeed: Colors.yellow,
   brightness: Brightness.dark,
 );
-
-Future<void> _initNotifications() async {
-  tz.initializeTimeZones();
-  final timeZoneName = await FlutterTimezone.getLocalTimezone();
-  tz.setLocalLocation(tz.getLocation(timeZoneName));
-
-  final notificationsPlugin = getIt<FlutterLocalNotificationsPlugin>();
-  const androidSettings = AndroidInitializationSettings('notification_icon');
-  const iosSettings = DarwinInitializationSettings(
-    requestAlertPermission: false,
-    requestBadgePermission: false,
-    requestSoundPermission: false,
-  );
-  const settings = InitializationSettings(
-    android: androidSettings,
-    iOS: iosSettings,
-  );
-  await notificationsPlugin.initialize(settings);
-}