+++ lib/service_locator.dart
@@ -4,6 +4,7 @@ import 'package:memorize_scripture/pages/home/home_page_manager.dart';
 import 'package:memorize_scripture/pages/practice/practice_page_manager.dart';
 import 'package:memorize_scripture/services/data_repository/data_repository.dart';
 import 'package:memorize_scripture/services/data_repository/sqflite/database.dart';
+import 'package:memorize_scripture/services/notification_service.dart';
 import 'package:memorize_scripture/services/user_settings.dart';
 import 'package:memorize_scripture/app_manager.dart';
 
@@ -12,8 +13,7 @@ final getIt = GetIt.instance;
 void setupServiceLocator() {
   getIt.registerLazySingleton<UserSettings>(() => SharedPreferencesStorage());
   getIt.registerLazySingleton<DataRepository>(() => LocalStorage());
-  getIt.registerLazySingleton<FlutterLocalNotificationsPlugin>(
-      () => FlutterLocalNotificationsPlugin());
+  getIt.registerLazySingleton<NotificationService>(() => NotificationService());
   getIt.registerFactory<PracticePageManager>(() => PracticePageManager());
   getIt.registerLazySingleton<AppManager>(() => AppManager());
   getIt.registerLazySingleton<HomePageManager>(() => HomePageManager());